updates compose file, entrypoint and re-adds protect_from_forgery

+++ entrypoint.sh
@@ -14,10 +14,13 @@ if [[ "$RAILS_ENV" != "production" && "$RAILS_ENV" != "prod" ]]; then
   DB_NAME="${DB_NAME}-${RAILS_ENV}"
 fi
 
-# run get_secrets.rb to create a temp env var file, then set 'em, then delete it
-/usr/bin/set_secrets.rb
-eval $(cat /tmp/secrets.env | sed 's/^/export /')
-rm -f /tmp/secrets.env
+if [[ ! -z "$AWS_REGION" && -z "$DISABLE_AWS_SECRETS" ]]; then
+  # specific code for AWS secrets manager, if being used. runs get_secrets.rb
+  # to create a temp env var file, then sets envs, then deletes file
+  /usr/bin/set_secrets.rb
+  eval $(cat /tmp/secrets.env | sed 's/^/export /')
+  rm -f /tmp/secrets.env
+fi
 
 # then pass vars to psql
 if echo "\c $DB_NAME; \dt" | psql -h "$PGHOST" -U "$PGUSER" -d "$DB_NAME" | grep schema_migrations 2>&1 >/dev/null
